requirements.txt
apache-airflow
polars
pandas
psycopg2-binary
sqlalchemy